backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
ultralytics==8.0.196
torch==2.0.1
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0 